+++ CHANGELOG.md
@@ -12,6 +12,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Exceptions raised through `transaction_if_not_already`
   when it was called from inside an existing transaction
   will no longer invalidate the outer transaction's context.
+- Remove `attrs` dependency.
 
 ### Added
 - Added MariaDB and SQLite to the test matrix.